YoutubeSummarizer/Backend/requirements.txt
flask
python-dotenv
google-generativeai
flask-cors
youtube-transcript-api
BeautifulSoup4
pymongo
requests